feat: Data Import [Insert, Update] Flow
Fill up the Google SpreadSheet form & click on the "Trigger Import"
button or make a REST API Call to trigger a fetch. This API fetches the
current spreadsheet data, prepares a set of un-imported data based on
the auto-updating counter field and generates a "Data Import" record.

The Data Import record is linked to the worksheet row as 'last_import'.
Only the last record is linked. However, all Import records can be found
in the "Data Import List View". Failures and Success summaries can be
found there too.

If "Skip Failures" is checked in the Worksheet row, new Data Imports can
be initiated even when the previous Import ended in Partial Success or
Failure states. Else, previous run will have to be rectified in order to
continue Importing.

* Trigger Google Spreadsheets importing via API Call / Single Click
* Add import type UPSERT stub
* Maintain frequency description on SpreadSheet doc
* Maintain client connection as instance attribute
* "Skip Failures" option in worksheet settings